// src/app/shared/interfaces/poke-details.interface.ts
export interface IPokeDetails {
    name: string;
    abilities: any;
    base_experience: number;
    forms: any;
    game_indices: any[];
    height: number;
    held_items: any[];
    id: number;
    is_default: boolean;
    location_area_encounters: string;
    moves: any[];
    order: number;
    past_types: any[];
    species: any;
    sprites: any;
    stats: any[];
    types: any[];
    weight: number;
}